+++ src/main/java/duke/Duke.java
@@ -1,11 +1,7 @@
-import java.util.ArrayList;
+package duke;
+
 import java.util.Scanner;
-import java.io.File;
 import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.format.DateTimeFormatter;
-import java.time.LocalDate;
 
 public class Duke {
     private Storage storage;
