+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.task.Task;
+
 import java.io.IOException;
 import java.util.ArrayList;
 public class TaskList {
@@ -8,7 +12,7 @@ public class TaskList {
     }
 
     TaskList(ArrayList<String> lst) {
-        //parse task from string to task
+        //parse duke.task from string to duke.task
         this.lst = new ArrayList<Task>();
         for (int i = 0; i < lst.size(); i++) {
             this.lst.add(Task.parseTask(lst.get(i)));
@@ -21,11 +25,11 @@ public class TaskList {
 
     public void deleteTask(int index) throws DukeException {
         if (index > lst.size() - 1 || index == 0) {
-            throw new DukeException("There is no such task");
+            throw new DukeException("There is no such duke.task");
         }
         Task t = lst.get(index);
         this.lst.remove(index);
-        System.out.println("Noted. I've removed this task: \n" + t.formatTask() + "\nNow you have " + lst.size() + " tasks in the list.");
+        System.out.println("Noted. I've removed this duke.task: \n" + t.formatTask() + "\nNow you have " + lst.size() + " tasks in the list.");
     }
 
     public int size() {
@@ -44,7 +48,7 @@ public class TaskList {
         Task t = lst.get(index);
         t.markAsDone();
         lst.set(index, t);
-        System.out.println("Nice! I've marked this task as done:");
+        System.out.println("Nice! I've marked this duke.task as done:");
         System.out.println(lst.get(index).formatTask());
 
     }
@@ -53,7 +57,7 @@ public class TaskList {
         Task t = lst.get(index);
         t.unMark();
         lst.set(index, t);
-        System.out.println("OK, I've marked this task as not done yet:");
+        System.out.println("OK, I've marked this duke.task as not done yet:");
         System.out.println(lst.get(index).formatTask());
     }
