+++ text-ui-test/EXPECTED.TXT
@@ -1,7 +1,7 @@
-Hello! I'm Duke.
+Hello! I'm duke.Duke.
 What can I do for you?
 
-Got it. I've added this task: 
+Got it. I've added this duke.task:
 [T] [ ] borrow book
 Now you have 1 tasks in the list.
 Bye! Hope to see you again soon!
