+++ text-ui-test/input.txt
@@ -1,2 +1,1 @@
 todo borrow book
-bye
